Moved omnodeconnection.h from indexer/indexgraph to include/om
Added new methods desc_from_file(), desc_from_string(), and
sort_desc() to OmIndexerBuilder interface, and get_node_info().

Fixed a few warnings in node source.


git-svn-id: file:///home/xapian-svn/xapian/trunk@2676 6b2f1b92-a4f3-0310-943f-de3cbe41748b

+++ xapian-core/indexer/indexgraph/omregexfilternode.cc
@@ -85,7 +85,7 @@ class OmRegexFilterNode : public OmIndexerNode {
 	    OmIndexerMessage output(new OmIndexerData(
 				    std::vector<OmIndexerData>()));
 
-	    for (int i=0; i<input->get_vector_length(); ++i) {
+	    for (size_t i=0; i<input->get_vector_length(); ++i) {
 		std::string orig = input->get_element(i).get_string();
 		if (regex.matches(orig)) {
 		    output->append_element(OmIndexerData(orig));
